Add song: Johk - La reine de France
I finnally managed to put the chords of `Quelqu'un m'a dit' on these lyrics... And that is really fun. For english reader, this songs his about the french first lady (and his beloved husband).

Have fun playing it. I feel like there is no introduction to this song. Just sing and start to play the music with a regular struming.